requirements.txt
streamlit
newsapi-python
feedparser
python-dotenv
openai
google-generativeai
pandas
requests
fuzzywuzzy
python-levenshtein
